Remove Horizon operations datasource
