+++ src/datasource/horizon/index.ts
@@ -1,7 +1,6 @@
 export * from "./assets";
 export * from "./effects";
-export * from "./operations";
 export * from "./order_book";
-export * from "./payments";
+export * from "./pathfinding";
 export * from "./trades";
 export * from "./transactions";
