+++ src/datasource/types.ts
@@ -5,18 +5,6 @@ import { AssetCode } from "../model/asset_code";
 export type HorizonAccountFlag = "auth_required" | "auth_immutable" | "auth_revocable";
 
 export type HorizonAssetType = "native" | "credit_alphanum4" | "credit_alphanum12";
-export type HorizonOpType =
-  | "create_account"
-  | "payment"
-  | "path_payment"
-  | "manage_offer"
-  | "create_passive_offer"
-  | "set_options"
-  | "change_trust"
-  | "allow_trust"
-  | "account_merge"
-  | "manage_data"
-  | "bump_sequence";
 
 type HorizonEffectType =
   | "account_created"
@@ -43,128 +31,6 @@ type HorizonEffectType =
   | "data_removed"
   | "data_updated"
   | "sequence_bumped";
-
-interface IBaseOperationData {
-  id: string;
-  paging_token: any;
-  source_account: AccountID;
-  transaction_hash: string;
-  created_at: string;
-  type: HorizonOpType;
-  type_i: number;
-}
-
-export interface ICreateAccountOperationData {
-  account: string;
-  funder: string;
-  starting_balance: string;
-}
-
-export interface IPaymentOperationData extends IBaseOperationData {
-  from: AccountID;
-  to: AccountID;
-  asset_type: HorizonAssetType;
-  asset_code: AssetCode;
-  asset_issuer: AccountID;
-  amount: string;
-}
-
-export interface IPathPaymentOperationData extends IBaseOperationData {
-  from: AccountID;
-  to: AccountID;
-  asset_code: AssetCode;
-  asset_issuer: AccountID;
-  asset_type: HorizonAssetType;
-  amount: string;
-  source_asset_code: AssetCode;
-  source_asset_issuer: AccountID;
-  source_asset_type: string;
-  source_max: string;
-  source_amount: string;
-}
-
-export interface IManageOfferOperationData extends IBaseOperationData {
-  offer_id: number;
-  amount: string;
-  buying_asset_code: AssetCode;
-  buying_asset_issuer: AccountID;
-  buying_asset_type: HorizonAssetType;
-  price: string;
-  price_r: { n: number; d: number };
-  selling_asset_code: AssetCode;
-  selling_asset_issuer: AccountID;
-  selling_asset_type: HorizonAssetType;
-}
-
-export interface ICreatePassiveOfferOperationData extends IBaseOperationData {
-  amount: string;
-  price: string;
-  price_r: { n: number; d: number };
-  buying_asset_type: HorizonAssetType;
-  buying_asset_code: AssetCode;
-  buying_asset_issuer: AccountID;
-  selling_asset_type: HorizonAssetType;
-  selling_asset_code: AssetCode;
-  selling_asset_issuer: AccountID;
-}
-
-export interface ISetOptionsOperationData extends IBaseOperationData {
-  signer_key: AccountID;
-  signer_weight: number;
-  master_key_weight: number;
-  low_threshold: number;
-  med_threshold: number;
-  high_threshold: number;
-  home_domain: string;
-  inflation_dest: string;
-  set_flags: number[];
-  set_flags_s: HorizonAccountFlag[];
-  clear_flags: number[];
-  clear_flags_s: HorizonAccountFlag[];
-}
-
-export interface IChangeTrustOperationData extends IBaseOperationData {
-  asset_code: AssetCode;
-  asset_issuer: AccountID;
-  asset_type: string;
-  trustee: AccountID;
-  trustor: AccountID;
-  limit: string;
-}
-
-export interface IAllowTrustOperationData extends IBaseOperationData {
-  asset_code: AssetCode;
-  asset_issuer: AccountID;
-  asset_type: string;
-  authorize: boolean;
-  trustee: AccountID;
-  trustor: AccountID;
-}
-
-export interface IAccountMergeOperationData extends IBaseOperationData {
-  into: AccountID;
-}
-
-export interface IManageDataOperationData extends IBaseOperationData {
-  name: string;
-  value: string;
-}
-
-export interface IBumpSequenceOperationData extends IBaseOperationData {
-  bump_to: string;
-}
-
-export type IHorizonOperationData = IPaymentOperationData &
-  ISetOptionsOperationData &
-  IAccountMergeOperationData &
-  IAllowTrustOperationData &
-  IBumpSequenceOperationData &
-  IChangeTrustOperationData &
-  ICreateAccountOperationData &
-  IManageDataOperationData &
-  IManageOfferOperationData &
-  ICreatePassiveOfferOperationData &
-  IPathPaymentOperationData;
 
 export interface IHorizonTransactionData {
   id: string;
