+++ src/schema/ledgers.ts
@@ -28,8 +28,6 @@ export const typeDefs = gql`
     transactions(first: Int, last: Int, before: String, after: String): TransactionConnection
     "Operations that were performed in this ledger"
     operations(first: Int, last: Int, before: String, after: String, order: Order): OperationConnection
-    "Payment-like operations that were performed in this ledger"
-    payments(first: Int, last: Int, before: String, after: String): OperationConnection
     "A list of all effects that have been produced in this ledger"
     effects(first: Int, last: Int, before: String, after: String): EffectConnection
   }
